fix: happy path test
