+++ .github/workflows/nutshell_itest.yml
@@ -20,7 +20,7 @@ jobs:
       - name: Rust Cache
         uses: Swatinem/rust-cache@v2
       - name: Test
-        run: nix develop -i -L .#stable --command cargo test -p cdk-integration-tests --tests nutshell_mint
+        run: nix develop -i -L .#stable --command cargo test -p cdk-integration-tests --tests happy_path_mint_wallet
       - name: Show logs if tests fail
         if: failure()
         run: docker logs nutshell
